Troubleshooting
The code seems correct. I began my troubleshooting process by turning the "Less Secure" setting on my Google account to "On". The reason for this is that Google blocks less secure email sources. That fixed my first problem, but I am still running in to errors. There is an issue sending the email through the SMTP connection because I am getting an "Authentication Required" error. I am still in the troubleshooting process.